♻️ Improve: `records_leaderboard` module structure [#63]
- ✏️ Rename: to `record_leaderboard`
- 🚚 Move: to `param` module
- ✏️ Rename: `RecordsLeaderboardSearchCriteria` struct to `SearchCriteria`
