+++ src/client/param/mod.rs
@@ -3,5 +3,6 @@
 pub mod news_stream;
 pub mod pagination;
 pub mod record;
+pub mod record_leaderboard;
 pub mod search_user;
 pub mod user_leaderboard;
